Data.Quotient

Documentation

data a :/ e infixl 7

class Equiv e a where

Minimal complete definition

toEquivClass, fromEquivClass

Associated Types

type EquivClass e a

Methods

toEquivClass :: Proxy e -> a -> EquivClass e a

fromEquivClass :: Proxy e -> EquivClass e a -> a

normalizeEquivClass :: Proxy (a :/ e) -> EquivClass e a -> EquivClass e a

eqBy :: Proxy (a :/ e) -> EquivClass e a -> EquivClass e a -> Bool

neqBy :: Proxy (a :/ e) -> EquivClass e a -> EquivClass e a -> Bool

quotient :: forall e a. Equiv e a => a -> a :/ e

(//) :: forall e a proxy. Equiv e a => a -> proxy e -> a :/ e infixl 7

getCanonical :: forall e a. Equiv e a => (a :/ e) -> a

getEquivClass :: (a :/ e) -> EquivClass e a

withEquivClass :: forall e a. Equiv e a => EquivClass e a -> a :/ e

unsafeWithEquivClass :: EquivClass e a -> a :/ e

liftQ :: forall e a b. (Equiv e a, Equiv e b) => (a -> b) -> (a :/ e) -> b :/ e

liftQ2 :: forall e a b c. (Equiv e a, Equiv e b, Equiv e c) => (a -> b -> c) -> (a :/ e) -> (b :/ e) -> c :/ e

liftQ3 :: forall e a b c d. (Equiv e a, Equiv e b, Equiv e c, Equiv e d) => (a -> b -> c -> d) -> (a :/ e) -> (b :/ e) -> (c :/ e) -> d :/ e

liftQ4 :: forall e a b c d r. (Equiv e a, Equiv e b, Equiv e c, Equiv e d, Equiv e r) => (a -> b -> c -> d -> r) -> (a :/ e) -> (b :/ e) -> (c :/ e) -> (d :/ e) -> r :/ e

liftQClass :: forall e f a b. (Equiv e a, Equiv f b) => (EquivClass e a -> EquivClass f b) -> (a :/ e) -> b :/ f

liftQClass2 :: forall e f g a b c. (Equiv e a, Equiv f b, Equiv g c) => (EquivClass e a -> EquivClass f b -> EquivClass g c) -> (a :/ e) -> (b :/ f) -> c :/ g

liftQClass3 :: forall e f g h a b c d. (Equiv e a, Equiv f b, Equiv g c, Equiv h d) => (EquivClass e a -> EquivClass f b -> EquivClass g c -> EquivClass h d) -> (a :/ e) -> (b :/ f) -> (c :/ g) -> d :/ h

liftQClass4 :: forall e f g h j a b c d r. (Equiv e a, Equiv f b, Equiv g c, Equiv h d, Equiv j r) => (EquivClass e a -> EquivClass f b -> EquivClass g c -> EquivClass h d -> EquivClass j r) -> (a :/ e) -> (b :/ f) -> (c :/ g) -> (d :/ h) -> r :/ j